+++ rich/console.py
@@ -10,6 +10,7 @@ from dataclasses import dataclass, field, replace
 from datetime import datetime
 from functools import wraps
 from getpass import getpass
+from itertools import islice
 from time import monotonic
 from typing import (
     IO,
@@ -22,6 +23,7 @@ from typing import (
     NamedTuple,
     Optional,
     TextIO,
+    Tuple,
     Union,
     cast,
 )
@@ -30,7 +32,7 @@ from typing_extensions import Literal, Protocol, runtime_checkable
 
 from . import errors, themes
 from ._emoji_replace import _emoji_replace
-from ._log_render import LogRender
+from ._log_render import LogRender, FormatTimeCallable
 from .align import Align, AlignMethod
 from .color import ColorSystem
 from .control import Control
@@ -40,6 +42,7 @@ from .measure import Measurement, measure_renderables
 from .pager import Pager, SystemPager
 from .pretty import Pretty
 from .scope import render_scope
+from .screen import Screen
 from .segment import Segment
 from .style import Style, StyleType
 from .styled import Styled
@@ -49,6 +52,7 @@ from .theme import Theme, ThemeStack
 
 if TYPE_CHECKING:
     from ._windows import WindowsConsoleFeatures
+    from .live import Live
     from .status import Status
 
 WINDOWS = platform.system() == "Windows"
@@ -56,6 +60,13 @@ WINDOWS = platform.system() == "Windows"
 HighlighterType = Callable[[Union[str, "Text"]], "Text"]
 JustifyMethod = Literal["default", "left", "center", "right", "full"]
 OverflowMethod = Literal["fold", "crop", "ellipsis", "ignore"]
+
+
+class NoChange:
+    pass
+
+
+NO_CHANGE = NoChange()
 
 
 CONSOLE_HTML_FORMAT = """\
@@ -82,10 +93,21 @@ body {{
 _TERM_COLORS = {"256color": ColorSystem.EIGHT_BIT, "16color": ColorSystem.STANDARD}
 
 
+class ConsoleDimensions(NamedTuple):
+    """Size of the terminal."""
+
+    width: int
+    """The width of the console in 'cells'."""
+    height: int
+    """The height of the console in lines."""
+
+
 @dataclass
 class ConsoleOptions:
     """Options for __rich_console__ method."""
 
+    size: ConsoleDimensions
+    """Size of console."""
     legacy_windows: bool
     """legacy_windows: flag for legacy windows."""
     min_width: int
@@ -104,6 +126,8 @@ class ConsoleOptions:
     """Disable wrapping for text."""
     highlight: Optional[bool] = None
     """Highlight override for render_str."""
+    height: Optional[int] = None
+    """Height available, or None for no height limit."""
 
     @property
     def ascii_only(self) -> bool:
@@ -112,30 +136,45 @@ class ConsoleOptions:
 
     def update(
         self,
-        width: int = None,
-        min_width: int = None,
-        max_width: int = None,
-        justify: JustifyMethod = None,
-        overflow: OverflowMethod = None,
-        no_wrap: bool = None,
-        highlight: bool = None,
+        width: Union[int, NoChange] = NO_CHANGE,
+        min_width: Union[int, NoChange] = NO_CHANGE,
+        max_width: Union[int, NoChange] = NO_CHANGE,
+        justify: Union[Optional[JustifyMethod], NoChange] = NO_CHANGE,
+        overflow: Union[Optional[OverflowMethod], NoChange] = NO_CHANGE,
+        no_wrap: Union[Optional[bool], NoChange] = NO_CHANGE,
+        highlight: Union[Optional[bool], NoChange] = NO_CHANGE,
+        height: Union[Optional[int], NoChange] = NO_CHANGE,
     ) -> "ConsoleOptions":
         """Update values, return a copy."""
         options = replace(self)
-        if width is not None:
+        if not isinstance(width, NoChange):
             options.min_width = options.max_width = width
-        if min_width is not None:
+        if not isinstance(min_width, NoChange):
             options.min_width = min_width
-        if max_width is not None:
+        if not isinstance(max_width, NoChange):
             options.max_width = max_width
-        if justify is not None:
+        if not isinstance(justify, NoChange):
             options.justify = justify
-        if overflow is not None:
+        if not isinstance(overflow, NoChange):
             options.overflow = overflow
-        if no_wrap is not None:
+        if not isinstance(no_wrap, NoChange):
             options.no_wrap = no_wrap
-        if highlight is not None:
+        if not isinstance(highlight, NoChange):
             options.highlight = highlight
+        if not isinstance(height, NoChange):
+            options.height = height
+        return options
+
+    def update_width(self, width: int) -> "ConsoleOptions":
+        """Update just the width, return a copy.
+
+        Args:
+            width (int): New width (sets both min_width and max_width)
+
+        Returns:
+            ~ConsoleOptions: New console options instance
+        """
+        options = replace(self, min_width=width, max_width=width)
         return options
 
 
@@ -249,6 +288,46 @@ class PagerContext:
         self._console._exit_buffer()
 
 
+class ScreenContext:
+    """A context manager that enables an alternative screen. See :meth:`~rich.console.Console.screen` for usage."""
+
+    def __init__(
+        self, console: "Console", hide_cursor: bool, style: StyleType = ""
+    ) -> None:
+        self.console = console
+        self.hide_cursor = hide_cursor
+        self.screen = Screen(style=style)
+        self._changed = False
+
+    def update(
+        self, renderable: RenderableType = None, style: StyleType = None
+    ) -> None:
+        """Update the screen.
+
+        Args:
+            renderable (RenderableType, optional): Optional renderable to replace current renderable,
+                or None for no change. Defaults to None.
+            style: (Style, optional): Replacement style, or None for no change. Defaults to None.
+        """
+        if renderable is not None:
+            self.screen.renderable = renderable
+        if style is not None:
+            self.screen.style = style
+        self.console.print(self.screen, end="")
+
+    def __enter__(self) -> "ScreenContext":
+        self._changed = self.console.set_alt_screen(True)
+        if self._changed and self.hide_cursor:
+            self.console.show_cursor(False)
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
+        if self._changed:
+            self.console.set_alt_screen(False)
+            if self.hide_cursor:
+                self.console.show_cursor(True)
+
+
 class RenderGroup:
     """Takes a group of renderables and returns a renderable object that renders the group.
 
@@ -298,15 +377,6 @@ def render_group(fit: bool = True) -> Callable:
         return _replace
 
     return decorator
-
-
-class ConsoleDimensions(NamedTuple):
-    """Size of the terminal."""
-
-    width: int
-    """The width of the console in 'cells'."""
-    height: int
-    """The height of the console in lines."""
 
 
 def _is_jupyter() -> bool:  # pragma: no cover
@@ -395,10 +465,12 @@ class Console:
             either ``"standard"``, ``"256"`` or ``"truecolor"``. Leave as ``"auto"`` to autodetect.
         force_terminal (Optional[bool], optional): Enable/disable terminal control codes, or None to auto-detect terminal. Defaults to None.
         force_jupyter (Optional[bool], optional): Enable/disable Jupyter rendering, or None to auto-detect Jupyter. Defaults to None.
+        force_interactive (Optional[bool], optional): Enable/disable interactive mode, or None to auto detect. Defaults to None.
         soft_wrap (Optional[bool], optional): Set soft wrap default on print method. Defaults to False.
         theme (Theme, optional): An optional style theme object, or ``None`` for default theme.
         stderr (bool, optional): Use stderr rather than stdout if ``file`` is not specified. Defaults to False.
         file (IO, optional): A file object where the console should write to. Defaults to stdout.
+        quiet (bool, Optional): Boolean to suppress all output. Defaults to False.
         width (int, optional): The width of the terminal. Leave as default to auto-detect width.
         height (int, optional): The height of the terminal. Leave as default to auto-detect height.
         style (StyleType, optional): Style to apply to all output, or None for no style. Defaults to None.
@@ -411,7 +483,7 @@ class Console:
         highlight (bool, optional): Enable automatic highlighting. Defaults to True.
         log_time (bool, optional): Boolean to enable logging of time by :meth:`log` methods. Defaults to True.
         log_path (bool, optional): Boolean to enable the logging of the caller by :meth:`log`. Defaults to True.
-        log_time_format (str, optional): Log time format if ``log_time`` is enabled. Defaults to "[%X] ".
+        log_time_format (Union[str, TimeFormatterCallable], optional): If ``log_time`` is enabled, either string for strftime or callable that formats the time. Defaults to "[%X] ".
         highlighter (HighlighterType, optional): Default highlighter.
         legacy_windows (bool, optional): Enable legacy Windows mode, or ``None`` to auto detect. Defaults to ``None``.
         safe_box (bool, optional): Restrict box options that don't render on legacy Windows.
@@ -428,10 +500,12 @@ class Console:
         ] = "auto",
         force_terminal: bool = None,
         force_jupyter: bool = None,
+        force_interactive: bool = None,
         soft_wrap: bool = False,
         theme: Theme = None,
         stderr: bool = False,
         file: IO[str] = None,
+        quiet: bool = False,
         width: int = None,
         height: int = None,
         style: StyleType = None,
@@ -443,7 +517,7 @@ class Console:
         highlight: bool = True,
         log_time: bool = True,
         log_path: bool = True,
-        log_time_format: str = "[%X]",
+        log_time_format: Union[str, FormatTimeCallable] = "[%X]",
         highlighter: Optional["HighlighterType"] = ReprHighlighter(),
         legacy_windows: bool = None,
         safe_box: bool = True,
@@ -475,6 +549,7 @@ class Console:
         self._color_system: Optional[ColorSystem]
         self._force_terminal = force_terminal
         self._file = file
+        self.quiet = quiet
         self.stderr = stderr
 
         if color_system is None:
@@ -498,6 +573,11 @@ class Console:
         self.no_color = (
             no_color if no_color is not None else "NO_COLOR" in self._environ
         )
+        self.is_interactive = (
+            (self.is_terminal and not self.is_dumb_terminal)
+            if force_interactive is None
+            else force_interactive
+        )
 
         self._record_buffer_lock = threading.RLock()
         self._thread_locals = ConsoleThreadLocals(
@@ -505,6 +585,7 @@ class Console:
         )
         self._record_buffer: List[Segment] = []
         self._render_hooks: List[RenderHook] = []
+        self._live: Optional["Live"] = None
 
     def __repr__(self) -> str:
         return f"<console width={self.width} {str(self._color_system)}>"
@@ -572,6 +653,25 @@ class Console:
         """Leave buffer context, and render content if required."""
         self._buffer_index -= 1
         self._check_buffer()
+
+    def set_live(self, live: "Live") -> None:
+        """Set Live instance. Used by Live context manager.
+
+        Args:
+            live (Live): Live instance using this Console.
+
+        Raises:
+            errors.LiveError: If this Console has a Live context currently active.
+        """
+        with self._lock:
+            if self._live is not None:
+                raise errors.LiveError("Only one live display may be active at once")
+            self._live = live
+
+    def clear_live(self) -> None:
+        """Clear the Live instance."""
+        with self._lock:
+            self._live = None
 
     def push_render_hook(self, hook: RenderHook) -> None:
         """Add a new render hook to the stack.
@@ -688,6 +788,7 @@ class Console:
     def options(self) -> ConsoleOptions:
         """Get default console options."""
         return ConsoleOptions(
+            size=self.size,
             legacy_windows=self.legacy_windows,
             min_width=1,
             max_width=self.width,
@@ -739,6 +840,16 @@ class Console:
         """
         width, _ = self.size
         return width
+
+    @property
+    def height(self) -> int:
+        """Get the height of the console.
+
+        Returns:
+            int: The height (in lines) of the console.
+        """
+        _, height = self.size
+        return height
 
     def bell(self) -> None:
         """Play a 'bell' sound (if supported by the terminal)."""
@@ -838,7 +949,7 @@ class Console:
         )
         return status_renderable
 
-    def show_cursor(self, show: bool = True) -> None:
+    def show_cursor(self, show: bool = True) -> bool:
         """Show or hide the cursor.
 
         Args:
@@ -846,6 +957,42 @@ class Console:
         """
         if self.is_terminal and not self.legacy_windows:
             self.control("\033[?25h" if show else "\033[?25l")
+            return True
+        return False
+
+    def set_alt_screen(self, enable: bool = True) -> bool:
+        """Enables alternative screen mode.
+
+        Note, if you enable this mode, you should ensure that is disabled before
+        the application exits. See :meth:`~rich.Console.screen` for a context manager
+        that handles this for you.
+
+        Args:
+            enable (bool, optional): Enable (True) or disable (False) alternate screen. Defaults to True.
+
+        Returns:
+            bool: True if the control codes were written.
+
+        """
+        changed = False
+        if self.is_terminal and not self.legacy_windows:
+            self.control("\033[?1049h\033[H" if enable else "\033[?1049l")
+            changed = True
+        return changed
+
+    def screen(
+        self, hide_cursor: bool = True, style: StyleType = None
+    ) -> "ScreenContext":
+        """Context manager to enable and disable 'alternative screen' mode.
+
+        Args:
+            hide_cursor (bool, optional): Also hide the cursor. Defaults to False.
+            style (Style, optional): Optional style for screen. Defaults to None.
+
+        Returns:
+            ~ScreenContext: Context which enables alternate screen on enter, and disables it on exit.
+        """
+        return ScreenContext(self, hide_cursor=hide_cursor, style=style or "")
 
     def render(
         self, renderable: RenderableType, options: ConsoleOptions = None
@@ -914,6 +1061,7 @@ class Console:
             options (Optional[ConsoleOptions], optional): Console options, or None to use self.options. Default to ``None``.
             style (Style, optional): Optional style to apply to renderables. Defaults to ``None``.
             pad (bool, optional): Pad lines shorter than render width. Defaults to ``True``.
+            range (Optional[Tuple[int, int]], optional): Range of lines to render, or ``None`` for all line. Defaults to ``None``
 
         Returns:
             List[List[Segment]]: A list of lines, where a line is a list of Segment objects.
@@ -927,6 +1075,10 @@ class Console:
                 _rendered, render_options.max_width, include_new_lines=False, pad=pad
             )
         )
+        if render_options.height is not None:
+            lines = Segment.set_shape(
+                lines, render_options.max_width, render_options.height, style=style
+            )
         return lines
 
     def render_str(
@@ -1007,7 +1159,9 @@ class Console:
         except errors.StyleSyntaxError as error:
             if default is not None:
                 return self.get_style(default)
-            raise errors.MissingStyle(f"Failed to get style {name!r}; {error}")
+            raise errors.MissingStyle(
+                f"Failed to get style {name!r}; {error}"
+            ) from None
 
     def _collect_renderables(
         self,
@@ -1166,6 +1320,7 @@ class Console:
         markup: bool = None,
         highlight: bool = None,
         width: int = None,
+        height: int = None,
         crop: bool = True,
         soft_wrap: bool = None,
     ) -> None:
@@ -1215,9 +1370,11 @@ class Console:
             render_options = self.options.update(
                 justify="default",
                 overflow=overflow,
-                width=min(width, self.width) if width else None,
+                width=min(width, self.width) if width else NO_CHANGE,
+                height=height,
                 no_wrap=no_wrap,
             )
+
             new_segments: List[Segment] = []
             extend = new_segments.extend
             render = self.render
@@ -1356,6 +1513,9 @@ class Console:
 
     def _check_buffer(self) -> None:
         """Check if the buffer may be rendered."""
+        if self.quiet:
+            del self._buffer[:]
+            return
         with self._lock:
             if self._buffer_index == 0:
                 if self.is_jupyter:  # pragma: no cover