+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.11.0] - 2021-02-15
+
+### Fixed
+
+- Fixed error message for tracebacks with broken `__str__` https://github.com/willmcgugan/rich/issues/980
+- Fixed markup edge case https://github.com/willmcgugan/rich/issues/987
+
+### Added
+
+- Added cheeky sponsorship request to test card
+- Added `quiet` argument to Console constructor
+- Added support for a callback function to format timestamps (allows presentation of milliseconds)
+- Added Console.set_alt_screen and Console.screen
+- Added height to ConsoleOptions
+- Added `vertical` parameter to Align
+- Added Layout class
+
+### Changed
+
+- Pretty.overflow now defaults to None
+- Panel now respects options.height
+- Traceback lexer defaults to Python if no extension on source
+- Added ConsoleDimensions size attribute to ConsoleOptions so that size can't change mid-render
+
 ## [9.10.0] - 2021-01-27
 
 ### Changed
@@ -27,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix double line tree guides on Windows
 - Fixed Tracebacks ignoring initial blank lines
 - Partial fix for tracebacks not finding source after chdir
+- Fixed error message when code in tracebacks doesn't have an extension https://github.com/willmcgugan/rich/issues/996
 
 ### Added